0x41d0c0: nOk in BOTH arms + a corrected param TYPE — 117 B -> 33 B at the exact extent
WorldgenPlaceItemForLockChainMaybe 0x41d0c0 goes 117 B @ len +13 -> 33 B @ len 401 =
Ghidra's extent EXACTLY, on two independent structural fixes.  v127 decomposed this
residual correctly at the level of MECHANISM and named both of its causes wrongly.

1. `nOk` must be assigned in BOTH ARMS of an if/else, not initialised and then
   conditionally overwritten.  `int nOk = 0; if (c) nOk = f();` starts nOk's live range
   before item1a's last use (`push ebx`), so the ranges overlap and nOk is homed to a
   frame slot — that spill IS the `sub esp,8` vs the original's `sub esp,4`.  With both
   arms assigning, cl sinks `mov ebx,0` INTO the WorldgenPlaceItemOnLock argument setup
   and re-uses item1a's now-dead register, exactly as the original does.  113 -> 33.
   The arm order is inert (3 spellings all 33/401); `!= 0`-first is pinned by the
   original's fallthrough (lesson #47).

2. WorldgenPlaceItemOnLock 0x41cdc0's 3rd parameter is `short`, not `int`.  v127
   dismissed this widthscan hit because "the callee reads that slot as a DWORD at four
   sites" — but both reads are FORWARDING PUSHES into another short parameter, and a
   forwarding push is byte-identical for `int` and `short`, so it discriminates nothing.
   The real proof runs the other way: PlaceQuestNode 0x41f120 loads its own nOrder with
   `mov bx, word [ebp+0x1c]` (garbage upper half) and pushes EBX RAW into 0x41d0c0,
   which pushes it RAW again into 0x41cdc0 — neither side widens.  FREE (+0/-0).

USER-APPROVED TRADE, 258 -> 257: fix 1 rotates the Worldgen.cpp joint phase and costs
SetCurrentToIntroZone 0x423d20 its exactness.  That function is a phase weathervane, not
a defect — its own floor was re-swept at the new phase (6 configurations, dead flat at 5,
countdown refuted by LENGTH) and its residual is now the SAME ESI<->EDI bijection
0x41d0c0 carries, so the two may fall together to a later TU-joint pass.

Both bodies' decl axes re-swept at the new phase and recorded as CLOSED in their notes.

New tool: tools/sbs.py — a side-by-side disassembler over residuals.paired(), printing
EVERY instruction on both sides in a range.  asmscore --dump prints only the DIFFERING
lines, which is unreadable once the schedule shifts; this view is what decomposed the
function.

Oracles: 257 exact / 99.17 % / link 0-0-exit0 / bugscan 1 HIGH (known benign) 0 SHIFT /
vt 10 CLEAN / msg 11 CLEAN / arity 0 mismatches (264 comparable).

+++ src/Worldgen.h
@@ -227,13 +227,13 @@ public:
                                                             // caller pushes it un-widened => short)
     int  WorldgenPlaceUsefulDropChainMaybe(short zoneId, short idx,
                                            short nOrder, short sel); // 0x0041cbe0
-    int  WorldgenPlaceItemOnLock(short zoneId, int a2, int nVal,
+    int  WorldgenPlaceItemOnLock(short zoneId, int a2, short nVal,
                                  short itemId, int sel);  // 0x0041cdc0
     int  WorldgenFillQuestItemSpot2Maybe(short zoneId, short a2, short nVal,
                                          unsigned short itemId);  // 0x0041cf10
     int  WorldgenPlaceItemForLockChainMaybe(short zoneId, short idx,
-                                            short nOrder, short sel); // 0x0041d0c0 (nOrder unread
-                                                            // in body; caller pushes un-widened)
+                                            short nOrder, short sel); // 0x0041d0c0 (nOrder is
+                                                    // pushed RAW into 0x41cdc0 - so nVal is short)
     int  WorldgenPlaceUsefulObjectMaybe(short zoneId, short itemId,
                                         short nOrder);       // 0x0041d260
     int  WorldgenAssignTransitItemMaybe(short zoneId, short nOrder,
